requirements.txt
Flask
python-dotenv
google-generativeai
google-api-python-client
